mejoras

+++ SANTA3/variables.py
@@ -11,7 +11,7 @@ kucoin_secret='409d3eff-9622-4488-af21-fa0feabb24ec'
 kucoin_passphares='santakucoin'
 
 if exchange_name == 'kucoin':
-    minvolumen24h = float(50000000)
+    minvolumen24h = float(100000000)
 else:
     minvolumen24h = float(200000000)
